chore: make client.py
